requirements.txt

Flask-SQLAlchemy
Flask-WTF
WTForms

dataset


